Simplify and refactor.
