+++ class-wc-facebookcommerce.php
@@ -136,7 +136,6 @@ if ( ! class_exists( 'WC_Facebookcommerce' ) ) :
 				require_once $this->get_framework_path() . '/utilities/class-sv-wp-background-job-handler.php';
 
 				require_once __DIR__ . '/includes/Products/class-fb-feed-generator.php';
-				require_once __DIR__ . '/includes/fbproductfeed.php';
 				require_once __DIR__ . '/facebook-commerce-messenger-chat.php';
 				require_once __DIR__ . '/includes/Exceptions/ConnectWCAPIException.php';
