+++ facebook-commerce.php
@@ -1697,70 +1697,6 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 		wp_die();
 	}
 
-
-	/**
-	 * Check Feed Upload Status (FBE v2.0)
-	 * TODO: When migrating to FBE v2.0, remove above function and rename
-	 * below function to ajax_check_feed_upload_status()
-	 **/
-	public function ajax_check_feed_upload_status_v2() {
-
-		\WC_Facebookcommerce_Utils::check_woo_ajax_permissions( 'check feed upload status', true );
-
-		check_ajax_referer( 'wc_facebook_settings_jsx' );
-
-		if ( $this->is_configured() ) {
-
-			$response = array(
-				'connected' => true,
-				'status'    => 'in progress',
-			);
-
-			if ( ! empty( $this->get_upload_id() ) ) {
-
-				if ( ! isset( $this->fbproductfeed ) ) {
-
-					if ( ! class_exists( 'WC_Facebook_Product_Feed' ) ) {
-						include_once 'includes/fbproductfeed.php';
-					}
-
-					$this->fbproductfeed = new \WC_Facebook_Product_Feed(
-						$this->get_product_catalog_id(),
-						$this->fbgraph
-					);
-				}
-
-				$status = $this->fbproductfeed->is_upload_complete( $this->settings );
-
-				$response['status'] = $status;
-
-			} else {
-
-				$response = array(
-					'connected' => true,
-					'status'    => 'error',
-				);
-			}
-
-			if ( 'complete' === $response['status'] ) {
-
-				update_option(
-					$this->get_option_key(),
-					apply_filters(
-						'woocommerce_settings_api_sanitized_fields_' . $this->id,
-						$this->settings
-					)
-				);
-			}
-		} else {
-
-			$response = array( 'connected' => false );
-		}
-
-		printf( json_encode( $response ) );
-		wp_die();
-	}
-
 	/**
 	 * Display custom success message (sugar)
 	 **/
