+++ includes/Products/FBCategories.php
@@ -15,8 +15,6 @@ defined( 'ABSPATH' ) || exit;
 /**
  * The main product feed handler.
  *
- * This will eventually replace \WC_Facebook_Product_Feed as we refactor and move its functionality here.
- *
  * @since 1.11.0
  */
 class FBCategories {
